+++ Result.py
@@ -1,0 +1,4 @@
+class Result:
+  def __init__(self, title, url):
+    self.title = title
+    self.url = url
