+++ University.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class University(Enum):
+  UCR="UCR"
+  ITCR="ITCR"
+  UNA="UNA"
+  UNED="UNED"
+  UTN="UTN"
